fix(rocr): size WaitMultiple event_age buffer with sizeof(uint64_t)
WaitEventAgeBytes now returns unique_evts * sizeof(uint64_t), and
Signal::WaitMultiple allocates via the helper instead of the wrong
sizeof(unique_evts) (sizeof(uint32_t)). The _alloca allocation now matches the
memset and the uint64_t element writes, so the stack buffer is no longer
under-allocated. Test passes.

+++ projects/rocr-runtime/runtime/hsa-runtime/core/inc/wait_event_age.h
@@ -55,7 +55,7 @@ namespace core {
 // the buffer must reserve sizeof(uint64_t) per unique signal. Centralising the
 // size here keeps the allocation and the memset from disagreeing.
 inline constexpr std::size_t WaitEventAgeBytes(uint32_t unique_evts) {
-  return static_cast<std::size_t>(unique_evts) * sizeof(uint32_t);
+  return static_cast<std::size_t>(unique_evts) * sizeof(uint64_t);
 }
 
 }  // namespace core
